vim-patch:8.2.3139: functions for string manipulation are spread out
Problem:    Functions for string manipulation are spread out.
Solution:   Move string related functions to a new source file. (Yegappan
            Lakshmanan, closes vim/vim#8470)

https://github.com/vim/vim/commit/a2438132a675be4dde3acbdf03ba1fdb2f09427c
