+++ src/nvim/generators/gen_eval.lua
@@ -35,6 +35,7 @@ hashpipe:write([[
 #include "nvim/quickfix.h"
 #include "nvim/runtime.h"
 #include "nvim/search.h"
+#include "nvim/strings.h"
 #include "nvim/sign.h"
 #include "nvim/testing.h"
